backend/foodgram/requirements.txt
asgiref==3.7.1
Brotli==1.0.9
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.1
cssselect2==0.7.0
defusedxml==0.7.1
Django==3.2.19
django-filter==23.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
docker==6.1.3
drf-base64==2.0
fonttools==4.38.0
gobject==0.1.0
html5lib==1.1
idna==3.4
importlib-metadata==1.7.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mypy==1.3.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
Pillow==9.5.0
psycopg2-binary==2.9.6
pycparser==2.21
pydyf==0.6.0
PyJWT==2.7.0
pyphen==0.14.0
python-dotenv==0.21.1
python3-openid==3.2.0
pytz==2023.3
pywin32==306
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
sqlparse==0.4.4
tinycss2==1.2.1
tomli==2.0.1
typed-ast==1.5.4
typing-extensions==4.6.2
uritemplate==4.1.1
urllib3==2.0.2
weasyprint==59.0
webencodings==0.5.1
websocket-client==1.5.2
zipp==3.15.0
zopfli==0.2.2
